좌석 상태 업데이트 로직 개선, 자료형 불일치 문제 해결

See merge request s09-final/S09P31D209!40

+++ tingting/src/main/java/com/alsif/tingting/concert/repository/ConcertSeatInfoRepository.java
@@ -11,7 +11,7 @@ import com.alsif.tingting.concert.dto.concerthall.SeatBookBaseDto;
 import com.alsif.tingting.concert.entity.ConcertSeatInfo;
 
 @Repository
-public interface ConcertSeatInfoRepository extends JpaRepository<ConcertSeatInfo, Integer> {
+public interface ConcertSeatInfoRepository extends JpaRepository<ConcertSeatInfo, Long> {
 
 	@Query("SELECT NEW com.alsif.tingting.concert.dto.concerthall.SeatBookBaseDto(csi.book) "
 		+ "FROM ConcertSeatInfo csi "
